requirements/base.txt
# Fabric
Fabric

# Django stuff
Django>=1.7.2

# Database Adapter
MySQL-python

# Templates
django_compressor

# Admin
django-debug-toolbar
django-docs

# utils
path.py
django-extensions
dj-database-url

# rest api
djangorestframework
markdown
django-filter
docutils

# angular
django-angular

# LSST
tornado
jinja2
SQLAlchemy
